\name{gsdesign-internal}
\title{Group Sequential Designs Internal Functions}
\alias{gsd.drift}
\description{
 Internal gsdesign functions.
}
\usage{
gsd.drift(ifrac, sig.level = 0.05, pow = 0.8, delta.eb = 0.5, delta.fb = 0, alternative = c("two.sided", "one.sided"), tol=0.0001)
}
\details{
  These functions are not to be called by the user.
}
\keyword{internal}
