\name{roc.perm.test-internal}
\title{Group Sequential Designs Internal Functions}
\alias{compareROC}
\description{
 Internal functions for permutation tests to compare ROC curves.
}
\usage{
  compareROC.paired(x,y,d,nperm=2500)
  compareROC.unpaired(x, dx, y, dy, nperm=2500, mp=NULL)
}
\details{
  These functions are not to be called by the user.
}
\keyword{internal}
