\name{ROCanalysis}
\alias{ROCanalysis}
\alias{rocanalysis}
\title{
  Functions to plot and compare ROC curves.
}
\description{
  These functions can be used for nonparametric analysis of ROC curves.
}
\details{
  The relevant functions are \code{roc.curve}, \code{roc.area.test} and
  \code{roc.perm.test}.  See the individual functions for usage details.
}
