% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-docs.R
\name{ph2single}
\alias{ph2single}
\title{Exact single stage Phase II design}
\usage{
ph2single(pu,pa,ep1,ep2,nsoln=5)
}
\arguments{
\item{pu}{unacceptable response rate; baseline response rate that needs to be exceeded for treatment to be deemed promising}

\item{pa}{response rate that is desirable; should be larger than pu}

\item{ep1}{threshold for the probability of declaring drug promising under pu (target type 1 error rate); between 0 and 1}

\item{ep2}{threshold for the probability of declaring the drug not promising under pa (target type 2 error rate); between 0 and 1}

\item{nsoln}{number of designs to be returned for given \code{ep1} and \code{ep2}}
}
\value{
ph2single returns a data frame with the following variables:
\itemize{
\item \code{n} - the total number of subjects treated (n)
\item \code{r} - the number of responses (r) needed to be exceeded to consider
treatment promising
\item \verb{Type I error}
\item \verb{Type II error}
}
}
\description{
Calculates the sample size and decision rule for exact single stage
Phase II design
}
\details{
The function assesses multiple designs and returns the \code{nsoln} possible
ones with the type-I and type-II errors closest to and below the target ones
given in arguments (\code{ep1}, \code{ep2}). The treatment is deemed promising if > \code{r}
responses are observed in the \code{n} patients.
}
\examples{
ph2single(0.2, 0.4, 0.1, 0.1)
ph2single(0.2, 0.35, 0.05, 0.05)
ph2single(0.2, 0.35, 0.05, 0.05, nsoln = 10)

}
\seealso{
Other phase2: 
\code{\link{ph2simon}()}
}
\concept{phase2}
\keyword{design}
